move builder to module
